1/requirements.txt

Flask
google-generativeai
Pillow
